Moved get_account_balance to TransactModel
