+++ models/transact_model.py
@@ -86,3 +86,14 @@ class TransactModel:
                 WHERE transactionid = %s
             """, (amount, transaction_id)
         )
+    @staticmethod
+    def get_account_balance(account_id):
+        """Calculate account balance using transaction table"""
+        result = db_fetchone("""
+                             SELECT COALESCE(SUM(amount), 0) as balance
+                             FROM transact
+                             WHERE accountid = %s
+                             """, (account_id,))['balance'] 
+        # db_fetchone() returns dict with only key 'balance'
+
+        return result
